Human Paths & Idle
Added human AI paths and filled map with a load of humans
Made a script to randomly play between a few idle animations for a few humans that are standing around.
