+++ Assets/Art/Low Poly Animated People/- Prefabs/grandma.prefab
@@ -157,6 +157,7 @@ GameObject:
   - component: {fileID: 54097394214157618}
   - component: {fileID: 65425903564767090}
   - component: {fileID: 114137608249409592}
+  - component: {fileID: 114855439647082994}
   m_Layer: 0
   m_Name: grandma
   m_TagString: Untagged
@@ -785,6 +786,19 @@ MonoBehaviour:
   startWaypointIndex: 0
   bloodSplatter: {fileID: 194062, guid: 8c27aeead6d968e4ba986702d6b84335, type: 2}
   scoreAwarded: 5
+--- !u!114 &114855439647082994
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1274296657605726}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9fe696d0a9c85544ea6e6d81d5850c36, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  distance: 5
+  throwingForce: 100
 --- !u!137 &137058688393905242
 SkinnedMeshRenderer:
   m_ObjectHideFlags: 1
