+++ Assets/Art/Low Poly Animated People/- Prefabs/kid.prefab
@@ -338,6 +338,7 @@ GameObject:
   - component: {fileID: 54035887154992582}
   - component: {fileID: 65799697945189708}
   - component: {fileID: 114279516028523060}
+  - component: {fileID: 114091269674605190}
   m_Layer: 0
   m_Name: kid
   m_TagString: Untagged
@@ -766,6 +767,19 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1, y: 2.100306, z: 1}
   m_Center: {x: 0, y: 1.0458401, z: 0}
+--- !u!114 &114091269674605190
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1829549139703218}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9fe696d0a9c85544ea6e6d81d5850c36, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  distance: 5
+  throwingForce: 100
 --- !u!114 &114279516028523060
 MonoBehaviour:
   m_ObjectHideFlags: 1
