+++ Assets/Art/Low Poly Animated People/- Prefabs/man.prefab
@@ -202,6 +202,7 @@ GameObject:
   - component: {fileID: 54326923616614688}
   - component: {fileID: 65790021463202088}
   - component: {fileID: 114754278264576922}
+  - component: {fileID: 114599641657491824}
   m_Layer: 0
   m_Name: man
   m_TagString: Untagged
@@ -766,6 +767,19 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1, y: 1.872544, z: 1}
   m_Center: {x: 0, y: 1.0487238, z: 0}
+--- !u!114 &114599641657491824
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1503856157617204}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9fe696d0a9c85544ea6e6d81d5850c36, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  distance: 5
+  throwingForce: 100
 --- !u!114 &114754278264576922
 MonoBehaviour:
   m_ObjectHideFlags: 1
