+++ Assets/Scripts/AI/Human.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1b139db394dc0d74a88a4ddbc7093990
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
