+++ Assets/Scripts/AI/Human/Human.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b2dc04f59a041ca40ad6003d6475f49d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
