+++ Assets/Scripts/AI/Human/IdleHuman.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8eccae964dde004abe828758eb360b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
