+++ ProjectSettings/TagManager.asset
@@ -24,6 +24,7 @@ TagManager:
   - Waypoint
   - Junction
   - TrafficLight
+  - FxTemporaire
   layers:
   - Default
   - TransparentFX
